finished

+++ src/routes/memories.ts
@@ -3,8 +3,15 @@ import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
+  app.addHook('preHandler', async (req) => {
+    await req.jwtVerify()
+  })
+
   app.get('/memories', async (req, rep) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -32,8 +39,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
+    if (!memory.isPublic && memory.userId !== req.user.sub) {
+      return rep.status(401).send()
+    }
+
     return memory
   })
+
   app.post('/memories', async (req, rep) => {
     const bodySchema = z.object({
       content: z.string(),
@@ -48,14 +60,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'f3d812f1-f19e-4222-99e1-02f165f12259',
+        userId: req.user.sub,
       },
     })
 
     return newMemory
   })
 
-  
   app.put('/memories/:id', async (req, rep) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
@@ -70,6 +81,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
 
     const { content, isPublic, coverUrl } = bodySchema.parse(req.body)
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== req.user.sub) {
+      return rep.status(401).send()
+    }
 
     const memoryUpdate = await prisma.memory.update({
       where: {
@@ -92,6 +113,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
 
     const { id } = paramsSchema.parse(req.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== req.user.sub) {
+      return rep.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
